Added configuration changes in the FE
