+++ project-timesheet-client/package.json
@@ -4,7 +4,7 @@
   "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --host www.onetimesheet.com --proxy-config config.json",
+    "start": "ng serve --host www.onetimesheet.com",
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
