+++ project-timesheet-client/src/app/app.component.html
@@ -17,11 +17,11 @@
 				Remember me
 			</label>
 		</div>
-		<button id="sign-in" class="btn btn-lg btn-primary btn-block"  type="submit">Sign
-			in</button>
-		<button id="tool-tip-bottom" class="btn btn-lg btn-primary btn-block " placement="bottom"
-			ngbTooltip="Not a user? please sign-up" type="submit">Sign
-			up</button>
+		<button id="sign-in" class="btn btn-lg btn-primary btn-block"
+			type="submit">Sign in</button>
+		<button id="tool-tip-bottom" class="btn btn-lg btn-primary btn-block "
+			placement="bottom" ngbTooltip="Not a user? please sign-up"
+			type="submit">Sign up</button>
 	</form>
 </div>
 <!-- /container -->
