+++ project-timesheet-client/src/app/services/configuration/configuration.service.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed, inject } from '@angular/core/testing';
+
+import { ConfigurationService } from './configuration.service';
+
+describe('ConfigurationService', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [ConfigurationService]
+    });
+  });
+
+  it('should be created', inject([ConfigurationService], (service: ConfigurationService) => {
+    expect(service).toBeTruthy();
+  }));
+});
