+++ project-timesheet-client/src/environments/environment.ts
@@ -3,6 +3,7 @@
 // `ng build --env=prod` then `environment.prod.ts` will be used instead.
 // The list of which env maps to which file can be found in `.angular-cli.json`.
 
- export const environment = {
-  production: false
- };
+export const environment = {
+  production: false,
+  configFile: 'assets/config/config.json'
+};
